implementar serviço irrf e melhorar classes de faixa de desconto

+++ src/main/java/br/folha/servico/INSSServico.java
@@ -140,7 +140,7 @@ public class INSSServico implements IINSSServico {
 		
 		FaixaINSSModel faixa = EncontrarFaixaV3(salarioBruto);
 		
-		MemoriaCalculoItem item = new MemoriaCalculoItem("INSS", faixa.getAliquota());
+		MemoriaCalculoItem item = new MemoriaCalculoItem("INSS", (float)faixa.getAliquota());
 	
 		item.setValor(faixa.obterValorDesconto(salarioBruto));
